requirements.txt
pyrtlsdr
gpsd-py3
requests
python-dotenv
geopandas
fiona
shapely
scapy
bleak
aiohttp
